Big rework: simplify explainer function structure
Checked equivalence in TF, works.
